--- Harvest/Assets/XR.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8083b73b7fe99964da4da301511cf516
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
